Bound the operands, serialize the writer, defer the unlink
Eight ways a declaration was taken for a guarantee.

Two chunks can share a chunk_id, because the PRIMARY KEY is the builder's word
like the rest of a pack's DDL. `associateBy` kept whichever came last, so the
duplicate did not merely break runtime lookups: it shrank the validator's own
view, and the checks that would have caught the rest of the damage never saw the
missing rows. Refused before that map is built, for the same reason a missing
table is.

`2d2147483647` is grammatical under unbounded "positive integers" and overflows
a 32-bit outcome range to a negative maximum, at which point coverage accepts a
table with no rows. The range is Long now, and the grammar pins N <= 100 and
S <= 1000 -- a distribution is a convolution N times over S faces, so anything
larger is not a table anyone printed.

Lexical search applied supersession to the query's results rather than inside
the query, so withdrawn passages consumed the depth budget and a common term
matching more than fifty corrected chunks could push every surviving hit out of
the list -- a refusal on a query the pack can answer.

The dense collapse ordered ties by chunk id alone, but a contract group spans
packs and two packs commonly both hold a chunk 1. Equal-scoring hits inherited
the query's order, RRF gave them different rank contributions, and the priority
tie-break downstream never saw a tie.

Normalization folds an ellipsis to whitespace, so "...for wizards?" reached the
follow-up check as "for wizards?" -- no opener, no rewrite, and a query sent to
retrieval having lost the thing it was about. Read before folding now.

StateDb shared one connection with no serialization, and autoCommit and the open
transaction are connection-wide state: a background install overlapping a
document edit pulled that edit into the install's transaction and lost it on
rollback. One lock, reentrant, with nested transactions joining rather than
committing the outer one early. A negative user_version is refused too, instead
of indexing MIGRATIONS[-1] and aborting every startup thereafter.

Uninstall unlinked the file immediately, contrary to the reader-lifetime
contract. Unlinking a SQLite file out from under an open connection is not a
crash on every platform, which is worse than if it were. Packs are borrowed
under a lease now; the row goes when the user says so and the bytes go when the
last reader closes.

+++ pack/src/main/kotlin/dev/rpghelper/pack/ReferenceValidation.kt
@@ -274,8 +274,8 @@ private fun checkCoverage(
     sorted: List<Triple<Long, Long, Long>>,
     out: MutableList<Violation>,
 ) {
-    val low = expression.min.toLong()
-    val high = expression.max.toLong()
+    val low = expression.min
+    val high = expression.max
 
     val outside = sorted.filter { it.first < low || it.second > high }
     if (outside.isNotEmpty()) {
